refactor(doc): extract HTML→Markdown into html_md; keep generics literal
Move the Javadoc HTML→Markdown conversion out of `doc` (which handles KDoc/Javadoc
*tags*) into a focused `html_md` module, so the two concerns don't mix.

Fixes the issue raised in review: the converter stripped *any* unrecognized
`<…>` as a tag, which also deleted non-HTML angle brackets like `List<String>`
or `Map<K, V>` in doc prose. It now strips only recognized HTML elements
(`is_known_html_tag` allowlist) and emits anything else literally. Dependency-free
— a real HTML parser would still tokenize `<String>` as an unknown element, so the
tag-vs-generic decision is an allowlist either way.

Adds html_md unit tests incl. regressions for `List<String>`, `Map<K, V>`,
`Flow<List<Item>>`, `a < b`, and generics mixed with real `<p>`.
